Add app.rs to take out app struct logic.
